third_party/blink/web_tests/external/wpt/webcodecs/video-encoder-rescaling.https.any_h264_avc-expected.txt

This is a testharness.js-based test.
[FAIL] Scaling Image in Encoding from 64 x 64 to 128 x 128 Format: I420
  assert_approx_equals: Scaled Image differs too much! Scaling from 64 x 64 to 128 x 128 Format:I420 expected 0 +/- 0.05 but got 0.35520482281454246
[FAIL] Scaling Image in Encoding from 32 x 32 to 96 x 96 Format: I420
  assert_approx_equals: Scaled Image differs too much! Scaling from 32 x 32 to 96 x 96 Format:I420 expected 0 +/- 0.05 but got 0.3893183778140886
[FAIL] Scaling Image in Encoding from 64 x 32 to 128 x 64 Format: I420
  assert_approx_equals: Scaled Image differs too much! Scaling from 64 x 32 to 128 x 64 Format:I420 expected 0 +/- 0.05 but got 0.3551470588235294
[FAIL] Scaling Image in Encoding from 64 x 64 to 128 x 192 Format: I420
  assert_approx_equals: Scaled Image differs too much! Scaling from 64 x 64 to 128 x 192 Format:I420 expected 0 +/- 0.05 but got 0.37569210409858395
[FAIL] Scaling Image in Encoding from 128 x 192 to 64 x 64 Format: I420
  assert_approx_equals: Scaled Image differs too much! Scaling from 128 x 192 to 64 x 64 Format:I420 expected 0 +/- 0.05 but got 0.20954765369689543
Harness: the test ran to completion.
